Migration 22: set height to bigint

+++ storage/migrations/22_gasoutputs_height_code.go
@@ -6,7 +6,7 @@ import "github.com/go-pg/migrations/v8"
 
 func init() {
 	up := batch(`
-	ALTER TABLE public.derived_gas_outputs ADD COLUMN height int8;
+	ALTER TABLE public.derived_gas_outputs ADD COLUMN height bigint;
 	ALTER TABLE public.derived_gas_outputs ADD COLUMN actor_name text;
 `)
 	down := batch(`
